USB/RNDIS: "iSerial" not set [SHLK-1254]
FLIR uses a method to assign rndis ip number based on usb port
This method relies on that device does not have a serial number.

This temporary solution is to avoid setting strings/0x409/serialNumber at
all. The driver will then work as FLIR expects it,
but it would probably not pass conformance tests.
-> More work needed (in usb gadget driver code)