Add implementation of read_pixels to g2d_renderer
Renderers within weston all have a method to retrieve frames called
read_pixels. In the case of the g2d_renderer the method was simply
returning success without providing any pixels. This caused screenshots
to be empty.

Weston has a screenshot client called weston-screenshooter. This
client is added as a target in the Makefile. By default weston only
allows screenshots to be taken if called from within weston. This
check is removed.