Activated webserver (lighttpd) and dependencies [SHLK-716]
flir-image-sherlock image now adds components:
lighttpd         (web server)
lighttpd modules (additional web support)
haproxy          (needed packet proxy)
php              (needed for http resource access)

To activate resource access using http://<camip>/prod/res/...,
web files needs to be added to /FLIR/usr/www (in a kit installation)
To be able to build the needed flir_res.so (php_flir_res repo) for Sherlock,
a new SDK must be first be generated from content in this (yocto) commit
(php needs to be present in SDK)
The (new) SDK to be installed.
php_flir_res source needs a to be checked out from a code branch where
cmake "php7" package usage is replaced with "php"

Target that starts to use an OS from this commit might need a full
"factory default" for haproxy+lighttpd to work.
/etc/passwd and /etc/group default content is updated with
pseudo-users www-data and haproxy.
Updated passwd and group files in ext4 are obscured with overlay
versions (which will be the case in a target that had a previous life)

yocto php package considerations:
php default recipe in yocto2.5 is actually a php7.
php7 as a specific package copied from flir-yocto:yocto2.0 does not build
Parallell installation of php7 and php5 which we use in flir-yocto
(Evander and Bellatrix) for SDK is not possible as of now.

Also note the php_7.2.7.bbappend
Fixes a build error that appears when valgrind is installed on the build
computer
See for instance https://github.com/openembedded/meta-openembedded/issues/94
A fix (for meta-openembedded later than in yocto2.5) was introduced as:
https://github.com/openembedded/meta-openembedded/commit/840850e834d9ee3d51b0ea9fcd169448e46e9747
By the .bbappend we have a flir backport