SHLK-2066 Enable Data pin contact detection
* Disable interfering pulldown resistors on the usb data pins prior to
  starting the USBDCD charger detection sequence.

* Enable D+ High to fulfill Dead Battery Provision requirements only at
  the end of the sequence instead of prior to starting the sequence.

U-boot runs a charger detection sequence using the USBDCD module to
determine the charge current to apply if a usb cable is
connected. Interfering pulldown resistors and current sources were
enabled which led the module to conclude that the data pins were always
in contact and immediately proceeded to next step in the sequence. If a
user connected a usb cable in a slow manner the module would try to
determine charger type prior to data pin connection and thus fail.