[CIRRUS-34] flir-image (wip): Add limited flir-image for evco
This flir-image is an incomplete work-in-progress and includes only
packages that can actually be built with the current setup. All packages in
the original evco image (in the yocto2.0 build) that are currently not
provided by any recipe have been collected in the UNAVAILABLE_PACKAGES
variable. All these packages need to be further worked on.