[CIRRUS-52] kernel: Add CONFIG_GPIO_MXC to fix gpio issues
This fixes, among other things, the initialization of fvdk.

The IMX kernel 5.10 does not have GPIO_MXC set as default so it needs to be
added to the kernel configuration. It is defined in
arch/arm/configs/imx_v7_defconfig for 5.10.y but our defconfig is based on
imx_v7_defconfig from the 5.4.y kernel which does not include it. Other
changes in this file between the two kernels do not seem necessary to pull
in.