CIRRUS-211 bitbake: Avoid Pseudo generating error 134 on image builds
Sometimes bitbake fails during image creation due to Pseudo (the fakeroot
system used by bitbake) mixing up inodes for certain files. It seems as if
inodes are changed for files without Pseudo knowing about it. This results
in an "error 134" meaning that the build task was sent the ABORT signal.

The error normally (or always?) appear in the execution of the
flir_rev_update() function and in the handling of the manifest file in
.repo.

This appears to be a new kind of problem in bitbake that has started
occuring with Yocto 3.2. Big changes have been made to how Pseudo handle
things in this release. The solution appears to be to use a variable
PSEUDO_IGNORE_PATHS to tell Pseudo not to consider files under certain
paths. In this attempt to solve the problem we also avoid setting the mode
of the manifest file since that might affect things, just keeping the time
data.

So this change does two things:
- Add ${BSPDIR}/.repo to PSEUDO_IGNORE_PATHS to keep pseudo from caring
  about files in that directory.
- Copy the file manifest.xml from .repo with --preserve=timestamps instead
  of preserving also mode and owners. Timestamp should be what we need
  anyways.

For further information on this issue, see:
https://docs.yoctoproject.org/ref-manual/migration-3.2.html#pseudo-path-filtering-and-mismatch-behaviour
and https://bugzilla.yoctoproject.org/show_bug.cgi?id=14175