flir-image-* and base-files-recovery: Remove build warnings
1. Remove inherit of "distro_features_check" from all image recipes since
it is deprecated and its replacement "features_check" is not used.

2. Don't use MACHINE specific files for fstab and profile in
base-files-recovery since it leads to warning when building for another
MACHINE (such as evco) and the files are identical for all known MACHINEs
anyway. If we need to single out a MACHINE in the future, just add a
files/<MACHINE>/ directory and add the MACHINE-specific files there.