qtbase-native: Remove use of statx system call
On systems that do not have statx system call (docker version of Ubuntu
18.04 for instance) qtwayland-native fails to link due to statx being used
in one of the libs from qtbase-native.

This is a somewhat hacky solution that removes the use of statx system call
for all cases (host types), but it is identical to what is used for all
Android systems so we are not alone and the change should be safe.