[BC-65] autofs, udev-extraconf: Enable autofs without MSD and UVC support
Remove all support for MSD (mass storage device) and MSD-combined modes
since we are now using MTP instead. Update autofs.sh script to handle
mounting and unmounting correctly. This will need to be further updated, or
some old functionality moved to a new script, when we want UVC to work.
autofs.sh should be cleaned from old commented code then.

udev-extraconf is updated to similar functionality as in Yocto2.0. For now
the most urgent need of update was to remove mount.sh and related
functionality since it collided with autofs.