[BC-22] Add z-ordering to weston
This adds functionality to Weston to place applications
in specific layers. Which layer an applications is placed
in is determined by the wayland app_id, that the application
needs to set.

This change supports an overlay layer (above normal applications,
which facet ends up in) a progress layer (above the overlay layer,
which progressapp should end up in), and the background layer
(below normal applications which streamserver's waylandsink ends
up in).