[BC-20] Support overlay on fbdev w Weston
Adds support for the option "filtered-outputs" to weston.ini. With
this option enabled along with weston configured with two fbdev
devices, the second device will only render views in the overlay
layer.

To use this, add the following to the [core] section of weston.ini...

filtered-outputs=true

...and launch weston with two fbdevs:

/usr/bin/weston [...] --device=/dev/fb0,/dev/fb4

This has only been tested with the same resolution on both fbdevs.

The filtered-outputs option will force Weston to place the outputs on
top of eachother (ie, clone mode, same x-offset). Weston will render
normally onto the first output but only render views on the second
output if they are positioned in the OVERLAY layer.

There is no explicit support for dynamically enabling/disabling this
behavior.

The implementation takes into account that weston's support for
overlapping outputs is broken by only considering damage for one
output. See:
https://gitlab.freedesktop.org/wayland/weston/-/merge_requests/357
https://gitlab.freedesktop.org/wayland/weston/-/issues/494
https://gitlab.freedesktop.org/wayland/weston/-/issues/504

Note that this change is not generic at all. It basically handles a
very narrow use-case and nothing more. So expect things breaking if
deviating from it.