[BC-81] Rec symbol handling
Add support to weston to configure local alpha on fbdevs supporting
it and to position apps and windows properly. This allows us to
run weston with two fbdevs (fb0 and fb1), where fb1 is an overlay
with local alpha support containing a separate overlay that is not
captured when recording.