Update display_enable.sh script, add hdmi display enable commands
Enabling the HDMI output for some unknown reason, requires us to configure the hdmi framebuffer to a higher resolution before going
to the wanted resolution of 640x480.
This was earlier implemented as a ExecStartPre clause in the weston service, but now, It seems to be neccessary to do this a little later.
Shoehorning the call in between the weston service and the flirapp
service using a new, hdmi-service

This should still be considered as TEMPORARY HACK TO ENABLE HDMI
Consider solving issue EOWYN-3191