eowyn hdmi: reinitialize hdmi display on display change [EOWYN-5097]
quickfix to reenable hdmi on display change
on display change, the kernel changes the hdmi vertical sync frequency

Eowyn uses 640x480 @ 60 Hz on HDMI,
during unblank though, HDMI is reinitialized and ends up outputting 95 Hz VSYNC,
which we have noted that som monitors are unable to synchronize to.

Reinitializing the hdmi again when the display is changed to LCD and Viewfinder sets
the HDMI VSYN back to 60 Hz.