gadgetload: usbfn, Remove CABLE_STATE and USBFNDRV settings
/etc/syfs-links/usb2_control/state, not working on Gxxx
and the reusult of the data appears to be unused in the script, so
removing It!

USBFNDRV, the output of lsmod | grep -o '^g_[a-z_]*' renders no output
on either Gxxx or basecamp cameras, (it exists on T1k though),
however the result of the data appears to be unused in the script, so
removing It!

[EOWYN-3685]