site.conf: avoid bitbake connectivity test
FLIR IT limits network access on build servers
bitbake will normally perform a "internet connection check" in its
build process.
There is a bitbake variable that controls this; "CONNECTIVITY_CHECK_URIS"
It is normally set by poky, but is possible to override.
If set to "" (as we now do) we disable the connectivity check

FLIR jenkins build slaves do not have "internet access".
We do have a "premirror" that handles this (for external repos) since
earlier.

This actual connectivity check needs to be handles separately as
IT internet access rules has changed recently