rpmsg-bifrost: AGATHA-1051 trigger USB CC callback
USB CC values would not be recognized at boot. Resulting in charger
not correctly set if this is required.

To set CC every time a new negotiation is initialized, use a callback
to ask m4 for CC values instead of using the pmic interrupt.

Remove annoying fuel gauge low state of charge warnings