imx-gst1.0-plugin: Use auto nIntraQP instead of configured quant [BC-652]
Patched imx-gst1.0-plugin to fix blinking defect in RTSP streamed image.

When using bit rate control to vpu encoder, the nIntraQP
parameter affects the detail in I-frames. Since this parameter also
depends on desired rate, resolution, fps, etc it is better to leave
it to auto, also to avoid inheriting a default QP.
In yocto2.0, it looks like this parameter was set to auto.