kernel: [RENATA-3895] reset flash before probe
This fixes issue that we could not mount flash of module if
it is running in XiP before reset.

If the flash is in Performance Enhanced mode(XiP) before trying to
get the ID from the flash, we need to reset the flash to normal mode
before probing the flash.