wlanload/ec501: yocto3.3 version to connect to wi-fi network upon reboot
yocto3.3 wi-fi reconnect to existing network did not seem to work at
all.
If restarting connman a long time (>14s) after wlan driver load
we will get things working
(if not, wlan0 is there, but not reconnected)

Probably this is not an optimal solution, but it seems to work
Should we include wpa-supplicant in some way instead?