weston, ec702: Tune weston repaint-window [LEIA-1198]
Tuned repaint-window parameter in weston.ini for ec702.
The upscaling to 1024x768 makes the rendering take longer and
the default 7ms value is too short.

Results were that weston missed the rendering deadline and
we only got around 28fps.
With a repaint-window setting of 13ms, 30 fps is reached.
Tried 9, 11, 13, 14, 15, 20 ms. It seems that 12-14ms is best.