LEIA-1168 Delay probe of power-supply regulator for boe vf
When the boe display was changed to be the primary display,
the timing was messed up so that the power-supply was deferred
when probed.
This in turn meant that the HDMI was successfully probed first and
the whole nxp fb init got messed up. It was then tried to probe the
power-supply when going into suspend. This did not work since the
actual ioexpander had already been turned off when we tried to disable it.

The solution here is to do the probe of the power-supply regulator later
by using a delayed work.